requirements.txt
autoprop==0.0.3
debugtools==0.2.0
future==0.17.1
glooey==0.1.2
more-itertools==7.0.0
Pillow==5.4.1
pyglet==1.3.2
PyYAML==5.1
six==1.12.0
vecrec==0.3.0
